chore(Queries): move result table hook to file [YTFRONT-5766]
